# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Polar API
  version: 0.2.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
paths:
  /healthz:
    get:
      security: []
      responses:
        '200':
          description: ok
  /readyz:
    get:
      security: []
      responses:
        '200':
          description: Ready
        '503':
          description: Not ready
  /v1/capabilities:
    get:
      responses:
        '200':
          description: Target-aware service capabilities
  /v1/targets:
    get:
      responses:
        '200':
          description: Configured monitoring targets
  /v1/station/health:
    get:
      responses:
        '200':
          description: Service health
  /v1/readings/latest:
    get:
      responses:
        '200':
          description: Latest indoor readings
  /v1/readings:
    get:
      parameters:
        - in: query
          name: metric
          required: true
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: resolution
          schema:
            type: string
          description: Optional Go duration for downsampling (for example 5m, 1h)
      responses:
        '200':
          description: Historical indoor readings
  /v1/forecast/latest:
    get:
      parameters:
        - in: query
          name: target
          schema:
            type: string
      responses:
        '200':
          description: Compatibility alias for latest target weather forecast
  /v1/weather/current:
    get:
      parameters:
        - in: query
          name: target
          schema:
            type: string
      responses:
        '200':
          description: Latest target weather conditions
  /v1/weather/forecast:
    get:
      parameters:
        - in: query
          name: target
          schema:
            type: string
      responses:
        '200':
          description: Latest target weather forecast
  /v1/weather/alerts:
    get:
      parameters:
        - in: query
          name: target
          schema:
            type: string
      responses:
        '200':
          description: Active weather alerts for the target
  /v1/air-quality/current:
    get:
      parameters:
        - in: query
          name: target
          schema:
            type: string
      responses:
        '200':
          description: Latest target air quality snapshot
  /v1/air-quality/forecast:
    get:
      parameters:
        - in: query
          name: target
          schema:
            type: string
      responses:
        '200':
          description: Latest target air quality forecast
  /v1/climate/snapshot:
    get:
      parameters:
        - in: query
          name: target
          schema:
            type: string
      responses:
        '200':
          description: Combined dashboard snapshot for a target
  /v1/live:
    get:
      parameters:
        - in: query
          name: target
          schema:
            type: string
      responses:
        '101':
          description: WebSocket upgrade for live dashboard updates
  /v1/diagnostics/data-gaps:
    get:
      responses:
        '200':
          description: Data gaps
  /v1/audit/events:
    get:
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: type
          schema:
            type: string
      responses:
        '200':
          description: Audit events
  /v1/metrics:
    get:
      responses:
        '200':
          description: Structured operational metrics snapshot